hw_detector and trezor services
